requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
requests==2.31.0
anthropic==0.7.0
scikit-learn==1.3.2
plotly==5.18.0
dash==2.14.1
google-generativeai==0.8.6